refactor: merge layered_drgporep and zigzag
