+++ filecoin-proofs/src/parameters.rs
@@ -1,12 +1,10 @@
 use storage_proofs::drgporep::DrgParams;
 use storage_proofs::drgraph::{DefaultTreeHasher, BASE_DEGREE};
 use storage_proofs::hasher::PedersenHasher;
-use storage_proofs::layered_drgporep;
-use storage_proofs::layered_drgporep::LayerChallenges;
 use storage_proofs::proof::ProofScheme;
 use storage_proofs::rational_post;
 use storage_proofs::rational_post::RationalPoSt;
-use storage_proofs::zigzag_drgporep::ZigZagDrgPoRep;
+use storage_proofs::zigzag_drgporep::{self, LayerChallenges, ZigZagDrgPoRep};
 use storage_proofs::zigzag_graph::{ZigZagBucketGraph, EXP_DEGREE};
 
 use crate::constants::POREP_MINIMUM_CHALLENGES;
@@ -26,7 +24,7 @@ pub type PostPublicParams = rational_post::PublicParams;
 pub fn public_params(
     sector_bytes: PaddedBytesAmount,
     partitions: usize,
-) -> layered_drgporep::PublicParams<DefaultTreeHasher, ZigZagBucketGraph<DefaultTreeHasher>> {
+) -> zigzag_drgporep::PublicParams<DefaultTreeHasher, ZigZagBucketGraph<DefaultTreeHasher>> {
     ZigZagDrgPoRep::<DefaultTreeHasher>::setup(&setup_params(sector_bytes, partitions)).unwrap()
 }
 
@@ -46,16 +44,10 @@ pub fn post_setup_params(post_config: PoStConfig) -> PostSetupParams {
 pub fn setup_params(
     sector_bytes: PaddedBytesAmount,
     partitions: usize,
-) -> layered_drgporep::SetupParams {
+) -> zigzag_drgporep::SetupParams {
     let sector_bytes = usize::from(sector_bytes);
 
-    let challenges = select_challenges(
-        partitions,
-        POREP_MINIMUM_CHALLENGES,
-        LAYERS,
-        TAPER_LAYERS,
-        TAPER,
-    );
+    let challenges = select_challenges(partitions, POREP_MINIMUM_CHALLENGES, LAYERS);
 
     assert!(
         sector_bytes % 32 == 0,
@@ -63,7 +55,7 @@ pub fn setup_params(
         sector_bytes,
     );
     let nodes = sector_bytes / 32;
-    layered_drgporep::SetupParams {
+    zigzag_drgporep::SetupParams {
         drg: DrgParams {
             nodes,
             degree: BASE_DEGREE,
@@ -78,24 +70,12 @@ fn select_challenges(
     partitions: usize,
     minimum_total_challenges: usize,
     layers: usize,
-    taper_layers: usize,
-    taper: f64,
 ) -> LayerChallenges {
     let mut count = 1;
-    let mut guess = LayerChallenges::Tapered {
-        count,
-        layers,
-        taper,
-        taper_layers,
-    };
-    while partitions * guess.total_challenges() < minimum_total_challenges {
+    let mut guess = LayerChallenges::new_fixed(count, layers);
+    while partitions * guess.challenges() < minimum_total_challenges {
         count += 1;
-        guess = LayerChallenges::Tapered {
-            count,
-            layers,
-            taper,
-            taper_layers,
-        };
+        guess = LayerChallenges::new_fixed(count, layers);
     }
     guess
 }
@@ -110,20 +90,13 @@ mod tests {
     #[test]
     fn partition_layer_challenges_test() {
         let f = |partitions| {
-            select_challenges(
-                partitions,
-                POREP_MINIMUM_CHALLENGES,
-                LAYERS,
-                TAPER_LAYERS,
-                TAPER,
-            )
-            .all_challenges()
+            select_challenges(partitions, POREP_MINIMUM_CHALLENGES, LAYERS).challenges()
         };
         // Update to ensure all supported PoRepProofPartitions options are represented here.
-        assert_eq!(vec![1, 1, 2, 2], f(usize::from(PoRepProofPartitions(2))));
+        assert_eq!(1, f(usize::from(PoRepProofPartitions(2))));
 
-        assert_eq!(vec![3, 3, 4, 5], f(1));
-        assert_eq!(vec![1, 1, 2, 2], f(2));
-        assert_eq!(vec![1, 1, 1, 1], f(4));
+        assert_eq!(1, f(1));
+        assert_eq!(2, f(2));
+        assert_eq!(4, f(4));
     }
 }
